simpri parameters included; parameter file output cleaned up; dual phase 1 revisited; PFI uses only absolute tolerance for elementaryFtran; Numerical cleaned up; Stable additions extended for parameterization;

+++ src/simplex/pricing/dualdevexpricing.cpp
@@ -142,7 +142,7 @@ void DualDevexPricing::update(int incomingIndex,
     Vector::NonzeroIterator nonzIterEnd = incomingAlpha.endNonzero();
     for (; nonzIter != nonzIterEnd; nonzIter++) {
         index = nonzIter.getIndex();
-        if (unlikely(index == outgoingIndex)) {
+        if (unlikely((int)index == outgoingIndex)) {
             continue;
         }
         Numerical::Double actualWeight = m_weights[index];
